pyqir tests: skip if not installed (#846)
* skip pyqir tests if not installed

* follow pr guidelines

* revision

* importorskip fixes

* fix tests

* fix changelog PR link

* add back packaging
